Fix launch action not usable for configuration without run mode
If a launch configuration does not support the launch mode "run" and
the configuration was never started before using Runner View the views
launch icon will be created as disabled run icon.
Because this also disables the icons drop-down menu it is not possible
to select the available run modes.

With this change the icon will show the first available launch mode
if "run" is not available.